+++ src/Sentry.Unity/ScreenshotEventProcessor.cs
@@ -1,4 +1,5 @@
 using Sentry.Extensibility;
+using Sentry.Unity.Integrations;
 using UnityEngine;
 
 namespace Sentry.Unity;
@@ -6,10 +7,13 @@ namespace Sentry.Unity;
 public class ScreenshotEventProcessor : ISentryEventProcessorWithHint
 {
     private readonly SentryUnityOptions _options;
+    private readonly IApplication _application;
+    public ScreenshotEventProcessor(SentryUnityOptions sentryOptions) : this(sentryOptions, null) { }
 
-    public ScreenshotEventProcessor(SentryUnityOptions sentryOptions)
+    internal ScreenshotEventProcessor(SentryUnityOptions sentryOptions, IApplication? application)
     {
         _options = sentryOptions;
+        _application = application ?? ApplicationAdapter.Instance;
     }
 
     public SentryEvent? Process(SentryEvent @event)
@@ -21,11 +25,18 @@ public class ScreenshotEventProcessor : ISentryEventProcessorWithHint
     {
         if (!MainThreadData.IsMainThread())
         {
+            _options.DiagnosticLogger?.LogDebug("Screenshot capture skipped. Can't capture screenshots on other than the main thread.");
             return @event;
         }
 
         if (_options.BeforeCaptureScreenshotInternal?.Invoke() is not false)
         {
+            if (_application.IsEditor)
+            {
+                _options.DiagnosticLogger?.LogInfo("Screenshot capture skipped. Capturing screenshots it not supported in the Editor");
+                return @event;
+            }
+
             if (Screen.width == 0 || Screen.height == 0)
             {
                 _options.DiagnosticLogger?.LogWarning("Can't capture screenshots on a screen with a resolution of '{0}x{1}'.", Screen.width, Screen.height);
@@ -37,7 +48,7 @@ public class ScreenshotEventProcessor : ISentryEventProcessorWithHint
         }
         else
         {
-            _options.DiagnosticLogger?.LogInfo("Screenshot attachment skipped by BeforeAttachScreenshot callback.");
+            _options.DiagnosticLogger?.LogInfo("Screenshot capture skipped by BeforeAttachScreenshot callback.");
         }